+++ src/scripts/plugins/alert.js
@@ -23,10 +23,7 @@ const template =
   <ui-dialog-title v-if="options.title">{{ options.title }}</ui-dialog-title>
   <ui-dialog-content>{{ options.message }}</ui-dialog-content>
   <ui-dialog-actions>
-    <ui-button
-      raised
-      primary
-      @click.native="handleClick">
+    <ui-button primary @click.native="handleClick">
       {{ options.buttonText }}
     </ui-button>
   </ui-dialog-actions>